% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPI_Allgather.R
\name{mpirc_MPI_Allgather}
\alias{mpirc_MPI_Allgather}
\title{Wrapper for MPI_Allgather()}
\usage{
mpirc_MPI_Allgather(send_vec)
}
\arguments{
\item{send_vec}{A vector of integers or numerics to be sent.}
}
\value{
A vector of the data from all nodes in rank order.
}
\description{
Each node sends a vector of the same size to all others,
and all nodes collect all the data.
}
