% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPI_Allreduce.R
\name{mpirc_MPI_Allreduce}
\alias{mpirc_MPI_Allreduce}
\title{Wrapper for MPI_Allreduce()}
\usage{
mpirc_MPI_Allreduce(data, func)
}
\arguments{
\item{data}{A vector of integers or numerics to be sent}

\item{func}{The reduction function; a string which can currently be "SUM",
"PROD", "MIN" or "MAX".}
}
\value{
On the root node, a vector of the same size as the original, having
applied the reduction function to the vectors from each node. On all other nodes, NULL.
}
\description{
Each node sends a vector of the same size to all other nodes,
which each perform a reduction operation (such as sum or product) across each
element, resulting in a single vector of the same size as the original.
}
