% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPI_Alltoall.R
\name{mpirc_MPI_Alltoall}
\alias{mpirc_MPI_Alltoall}
\title{Wrapper for MPI_Alltoall()}
\usage{
mpirc_MPI_Alltoall(data, root)
}
\arguments{
\item{data}{A vector of integers or numerics to be sent}
}
\value{
A vector of the contributions of data received from each node in order of rank
is returned.
}
\description{
Each node chops its data into equally-sized pieces, one for each
node, and scatters the data to each. The receiving node gathers the equally-sized
pieces from each sending node. For example, if there are three nodes, and six 
elements of data, then node 0 sends the first and second elements to node 0 (itself), 
the third and fourth elements to node 1, and the fifth and sixth elements to node 2. 
After the call, node 0 has all the first and second elements from everybody (in node 
order); node 1 has all the third and fourth elements similarly, and node 2 has all 
the fifth and sixth. Hence, MPI_Alltoall is a little like a transpose, in the way
data gets distributed.
}
