% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPI_Barrier.R
\name{mpirc_MPI_Barrier}
\alias{mpirc_MPI_Barrier}
\title{MPI Barrier}
\usage{
mpirc_MPI_Barrier()
}
\description{
Wrapper for MPI_Barrier(world). Causes MPI processes to
block until all of the team reach the barrier, and only the continue.
Use this to ensure parallel work is completed before the next section
of code starts.
}
