% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPI_Bcast.R
\name{mpirc_MPI_Bcast}
\alias{mpirc_MPI_Bcast}
\title{Wrapper for MPI_Bcast()}
\usage{
mpirc_MPI_Bcast(data, root)
}
\arguments{
\item{data}{The vector of data to be sent}

\item{root}{The rank of the node that performs the broadcast}
}
\value{
The vector of data from the broadcasting node.
}
\description{
The root node sends identical data to all other nodes
}
