% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPI_Comm_rank.R
\name{mpirc_MPI_Comm_rank}
\alias{mpirc_MPI_Comm_rank}
\title{Wrapper for MPI_Comm_rank(MPI_COMM_WORLD, &my_rank)}
\usage{
mpirc_MPI_Comm_rank()
}
\value{
Rank of this node, between 0 and (size-1) inclusive.
}
\description{
Wrapper for MPI_Comm_rank(MPI_COMM_WORLD, &my_rank)
}
