% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPI_Comm_size.R
\name{mpirc_MPI_Comm_size}
\alias{mpirc_MPI_Comm_size}
\title{Wrapper for MPI_Com_size(MPI_COMM_WORLD, &no_nodes)}
\usage{
mpirc_MPI_Comm_size()
}
\value{
Number of nodes included in the MPI_COMM_WORLD.
}
\description{
Wrapper for MPI_Com_size(MPI_COMM_WORLD, &no_nodes)
}
