% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPI_Finalize.R
\name{mpirc_MPI_Finalize}
\alias{mpirc_MPI_Finalize}
\title{Finalize MPI communications.}
\usage{
mpirc_MPI_Finalize()
}
\description{
Wrapper for MPI_Finalize(). Call this only once at the end of
the code.
}
