% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPI_Gather.R
\name{mpirc_MPI_Gather}
\alias{mpirc_MPI_Gather}
\title{Wrapper for MPI_Gather()}
\usage{
mpirc_MPI_Gather(send_vec, root)
}
\arguments{
\item{send_vec}{A vector of integers or numerics to be sent.}
}
\value{
On the root node, a vector of the data from all nodes in order
is returned; on all other nodes, NULL.
}
\description{
Each node sends a vector of the same size to all others,
and the root node alone collects all the data.
}
