% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPI_Gatherv.R
\name{mpirc_MPI_Gatherv}
\alias{mpirc_MPI_Gatherv}
\title{Wrapper for MPI_Gatherv()}
\usage{
mpirc_MPI_Gatherv(data, counts, displacements, root)
}
\arguments{
\item{data}{A vector of integers or numerics to be sent}

\item{counts}{The size of the vector send by each node}

\item{displacements}{The position in the resulting vector where each
node's contributing data is put; the first element is zero.}

\item{root}{The rank of node that receives the gathered data}
}
\value{
On the root node, a vector of the data from all nodes, arranged
as specified by the counts and displacements. On all other nodes, NULL.
}
\description{
Each node sends a vector (of any size) to all others,
and the root node alone collects all the data into one vector,
with sizes and offsets in the final vector declared.
}
