% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPI_Get_processor_name.R
\name{mpirc_MPI_Get_processor_name}
\alias{mpirc_MPI_Get_processor_name}
\title{Get processor name}
\usage{
mpirc_MPI_Get_processor_name()
}
\value{
The name of the compute node calling the function.
}
\description{
Wrapper for MPI_Get_processor_name.
}
