% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPI_Init.R
\name{mpirc_MPI_Init}
\alias{mpirc_MPI_Init}
\title{Initialise MPI communcation}
\usage{
mpirc_MPI_Init()
}
\description{
Wrapper for MPI_Init(NULL, NULL). Call only once at the start
of the code. After this call, you can get the size of the MPI world and
this node's rank.
}
