% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPI_Init_thread.R
\name{mpirc_MPI_Init_thread}
\alias{mpirc_MPI_Init_thread}
\title{Initialise MPI communcation}
\usage{
mpirc_MPI_Init_thread(level)
}
\arguments{
\item{level}{The name of the threading level, as per the MPI documentation.
One of `MPI_THREAD_SINGLE` (which implies no threading), `MPI_THREAD_FUNELLED`
(in which only the master thread performs MPI functions), `MPI_THREAD_SERIALIZED`
(in which any threads can perform an MPI function, but not simultaneously), or
`MPI_THREAD_MULTIPLE`, where any threads can perform MPI functions at the same
time. Performance is reported to be worse in this last mode, while algorithms
using it will be somewhat complicated to synchronise. These four options are
numberd 0, 1, 2 and 3 internally - see return value.}
}
\value{
The level that the MPI implementation provides; you should check that
this number is greater or equal to the level you request.
}
\description{
Wrapper for MPI_Init_threads(). Call only once at the start
of the code, as an alternative to MPI_Init, specifying the level of threading
you want. MPI implementations are not required to support all the levels of
threading. For OpenMP use, using MPI_Init on a whole node, followed by
`omp_init()` is sufficient for hybrid MPI/OpenMP.
}
