% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPI_Initialized.R
\name{mpirc_MPI_Initialized}
\alias{mpirc_MPI_Initialized}
\title{Test if MPI is initialised}
\usage{
mpirc_MPI_Initialized()
}
\description{
Wrapper for MPI_Initialized(). Useful since other operations
will crash R if called from an inappropriate state.
}
