% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPI_Reduce.R
\name{mpirc_MPI_Reduce}
\alias{mpirc_MPI_Reduce}
\title{Wrapper for MPI_Reduce()}
\usage{
mpirc_MPI_Reduce(data, root, func)
}
\arguments{
\item{data}{A vector of integers or numerics to be sent}

\item{root}{The rank of node that receives the data}

\item{func}{The reduction function; a string which can currently be "SUM",
"PROD", "MIN" or "MAX".}
}
\value{
On the root node, a vector of the same size as the original, having
applied the reduction function to the vectors from each node. On all other nodes, NULL.
}
\description{
Each node sends a vector of the same size to the root node,
which performs a reduction operation (such as sum or product) across each
element, resulting in a single vector of the same size as the original.
}
