% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPI_Scatter.R
\name{mpirc_MPI_Scatter}
\alias{mpirc_MPI_Scatter}
\title{Wrapper for MPI_Scatter()}
\usage{
mpirc_MPI_Scatter(data, root)
}
\arguments{
\item{data}{A vector of integers or numerics.}

\item{root}{The node with the data to be scattered}
}
\value{
On each node, an equally sized chunk of the root node's data
}
\description{
The root node sends an equally-sized chunk of its data to
each node.
}
