% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPI_Wtime.R
\name{mpirc_MPI_Wtime}
\alias{mpirc_MPI_Wtime}
\title{Wrapper for MPI_Wtime()}
\usage{
mpirc_MPI_Wtime()
}
\value{
Returns a high-resolution time in seconds since some reference point.
}
\description{
Wrapper for MPI_Wtime()
}
